credits work

+++ cart.php
@@ -4,7 +4,19 @@ require_once 'login.php';
 $conn = new mysqli($hn, $un, $pw, $db);
 if ($conn->connect_error) die($conn->connect_error);
 $buyid= $_SESSION["tmpid"];
-echo $buyid;
+echo "your id: ". $buyid;
+ $total="";
+ $query = "SELECT * from customer where id='$buyid'";
+	$result = $conn->query($query);
+	while ($row = $result->fetch_assoc()) {
+			//echo "Welcome ".$tmp." , id: ".$row['id']."<br>";
+			$ti=$row['id'];
+			$tcustcre=$row['credits'];
+			}
+	if (!$result) echo "SELECT failed: $query<br>" . $conn->error . "<br><br>";
+	echo"</br>";
+	echo "you have $".$tcustcre;
+	echo "</br>";
 /*
 if (isset($_POST['delete']) && isset($_POST['id']))
 {
@@ -15,20 +27,24 @@ if (isset($_POST['delete']) && isset($_POST['id']))
 	if (!$result) echo "DELETE failed: $query<br>" . $conn->error . "<br><br>";
 	}
 	*/
+	
 
 //delete from cart
-if (isset($_POST['delete']) && isset($_POST['id']) && isset($_POST['res']) && isset($_POST['size']) && isset($_POST['source']) && isset($_POST['category']))
+if (isset($_POST['delete']) && isset($_POST['id']) && isset($_POST['res']) && isset($_POST['size']) && isset($_POST['source']) && isset($_POST['category']) && isset($_POST['credits']))
 {
 	echo "hi";
 	$id = get_post($conn, 'id');
 	$res = get_post($conn, 'res');
 	$size = get_post($conn, 'size');
 	$name = get_post($conn, 'source');
-		$cat = get_post($conn, 'category');
+	$cat = get_post($conn, 'category');
+	$cre = get_post($conn, 'credits');
 	
-	$query = "INSERT INTO music VALUES('$id','$res','$size','$name','$cat')";
+	$query = "INSERT INTO music VALUES('$id','$res','$size','$name','$cat','$cre')";
 	$result = $conn->query($query);
 	if (!$result) die("Database access failed: ". $conn->error);
+	
+	$total=$total-$cre;
 	
 	$query = "DELETE FROM cart WHERE id='$id'";
 	$result = $conn->query($query);
@@ -42,27 +58,38 @@ if (isset($_POST['delete']) && isset($_POST['id']) && isset($_POST['res']) && is
 	
 	$query ="SELECT * FROM cart";
 		$result = $conn->query($query);
+		$tcred=array();
 		
 		while ($row = $result->fetch_assoc()) {
 		$s[]=$row['id'];
-		print_r ($s);
+		//print_r ($s);
 		$ts=$row['source'];
 			$newts[]=substr_replace($ts, '.jpg',-7);
 		$tc[]=$row['category'];
-					
-		
-		
+		$tcred[]=$row['credits'];
 	}
+	print_r ($tcred);
+	$sum= array_sum($tcred);
 		if (!$result) echo "SELECT failed: $query<br>" . $conn->error . "<br><br>";
-	$query = "SELECT id from customer where id='$buyid'";
+	$query = "SELECT * from customer where id='$buyid'";
 	$result = $conn->query($query);
 	while ($row = $result->fetch_assoc()) {
 			//echo "Welcome ".$tmp." , id: ".$row['id']."<br>";
 			$ti=$row['id'];
+			$tcustcre=$row['credits'];
 			}
 	if (!$result) echo "SELECT failed: $query<br>" . $conn->error . "<br><br>";
+	echo "</br>";
+	echo $sum;
+	if($tcustcre > $sum)
+	{
+		$upcustcre=$tcustcre-$sum;
+		echo $upcustcre;
+		$query = "update customer set credits=$upcustcre where id='$buyid'";
+	$result = $conn->query($query);
+	if (!$result) echo "update failed: $query<br>" . $conn->error . "<br><br>";
 	
-	
+
 		foreach ($s as $k => $v) {
     $query ="INSERT INTO transaction VALUES(NULL,'$ti','$s[$k]','$newts[$k]','$tc[$k]',CURDATE())";
 	$result = $conn->query($query);
@@ -70,10 +97,17 @@ if (isset($_POST['delete']) && isset($_POST['id']) && isset($_POST['res']) && is
 			
 }
 
+
 	
 	$query = "DELETE FROM cart";
 	$result = $conn->query($query);
 	if (!$result) echo "DELETE failed: $query<br>" . $conn->error . "<br><br>";
+	
+	}
+	else{
+		echo "not enough credits bitch $$, delete few items you broke arse mfucker";
+		header("location: cart.php");
+	}
 	
 	}
 echo <<<_END
@@ -108,7 +142,14 @@ _END;
 	
 	$query = "SELECT * FROM cart";
 		$result = $conn->query($query);
+		$tcred=array();
+		while ($row = $result->fetch_assoc()) {
+		$tcred[]=$row['credits'];
+	}
+	$total=array_sum($tcred);
+	echo "total: ".$total;
 		if (!$result) die ("Database access failed: " . $conn->error);
+		
 		$rows = $result->num_rows;
 		
 		for ($j = 0 ; $j < $rows ; ++$j)
@@ -123,6 +164,7 @@ resolution $row[1]
 size $row[2]
 source $row[3]
 category: $row[4]
+credits: $row[5]
 <img src= $row[3] alt="HTML5 Icon" style="width:128px;height:128px">
 </pre>
 
@@ -133,6 +175,7 @@ category: $row[4]
 <input type="hidden" name="size" value="$row[2]">
 <input type="hidden" name="source" value="$row[3]">
 <input type="hidden" name="category" value="$row[4]">
+<input type="hidden" name="credits" value="$row[5]">
 
 <input type="submit" value="DELETE RECORD">
 </form>
